scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "AUR packages update tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "pathlib",
    "argparse>=1.4.0",
]
